+++ ceammc/ext/src/preset/preset_storage.h
@@ -14,6 +14,7 @@ public:
     AtomList p_keys() const;
 
     void m_load(t_symbol*, const AtomListView& l);
+    void m_interp(t_symbol*, const AtomListView& l);
     void m_store(t_symbol*, const AtomListView& l);
     void m_clear(t_symbol*, const AtomListView& l);
     void m_write(t_symbol*, const AtomListView& fname);